+++ pkg/kubectl/cmd/util/openapi/openapi_getter.go
@@ -29,8 +29,6 @@ type synchronizedOpenAPIGetter struct {
 	openAPISchema Resources
 	err           error
 
-	serverVersion string
-	cacheDir      string
 	openAPIClient discovery.OpenAPISchemaInterface
 }
 
@@ -42,12 +40,10 @@ type Getter interface {
 	Get() (Resources, error)
 }
 
-// NewOpenAPIGetter returns an object to return OpenAPIDatas which either read from a
-// local file cache or read from a server, and then stored in memory for subsequent invocations
-func NewOpenAPIGetter(cacheDir, serverVersion string, openAPIClient discovery.OpenAPISchemaInterface) Getter {
+// NewOpenAPIGetter returns an object to return OpenAPIDatas which reads
+// from a server, and then stores in memory for subsequent invocations
+func NewOpenAPIGetter(openAPIClient discovery.OpenAPISchemaInterface) Getter {
 	return &synchronizedOpenAPIGetter{
-		serverVersion: serverVersion,
-		cacheDir:      cacheDir,
 		openAPIClient: openAPIClient,
 	}
 }
@@ -55,15 +51,13 @@ func NewOpenAPIGetter(cacheDir, serverVersion string, openAPIClient discovery.Op
 // Resources implements Getter
 func (g *synchronizedOpenAPIGetter) Get() (Resources, error) {
 	g.Do(func() {
-		client := NewCachingOpenAPIClient(g.openAPIClient, g.serverVersion, g.cacheDir)
-		result, err := client.OpenAPIData()
+		s, err := g.openAPIClient.OpenAPISchema()
 		if err != nil {
 			g.err = err
 			return
 		}
 
-		// Save the result
-		g.openAPISchema = result
+		g.openAPISchema, g.err = NewOpenAPIData(s)
 	})
 
 	// Return the save result
